# apps/docs/src/data/openapi/nest-auth.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/mfa/challenge": {
      "post": {
        "operationId": "MfaController_challenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthSendMfaCodeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthMfaCodeSentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send MFA code for setup/verification",
        "tags": [
          "Mfa"
        ]
      }
    },
...
    "/auth/mfa/toggle": {
      "post": {
        "operationId": "MfaController_toggleMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthToggleMfaRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthMfaToggleResponseDto"
                }
              }
            }
          }
        },
        "summary": "Enable or disable MFA for the current user",
        "tags": [
          "Mfa"
        ]
      }
    },
...
    "/auth/mfa/devices/{deviceId}": {
      "delete": {
        "operationId": "MfaController_removeDevice",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthMfaDeviceRemovedResponseDto"
                }
              }
            }
          }
        },
        "summary": "Remove a registered MFA device",
        "tags": [
          "Mfa"
        ]
      }
    },
...
    "/auth/mfa/verify-totp-setup": {
      "post": {
        "operationId": "MfaController_verifyTotpSetup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthVerifyTotpSetupRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthMfaDeviceVerifiedResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify TOTP Setup",
        "tags": [
          "Mfa"
        ]
      }
    },
...
  },
  "info": {
    "title": "@ackplus/nest-auth API",
    "description": "OpenAPI specification generated from the Nest Auth module",
    "version": "2.0.0-beta.16",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NestAuthMfaCodeSentResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "MFA code sent successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "NestAuthMFAMethodEnum": {
        "type": "string",
        "enum": [
          "email",
          "sms",
          "totp"
        ]
      },
...
      "NestAuthToggleMfaRequestDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Whether MFA should be enabled for the current user",
            "example": true
          }
        },
        "required": [
          "enabled"
        ]
      },
      "NestAuthMfaToggleResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "MFA enabled successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "NestAuthMfaDeviceRemovedResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Device removed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "NestAuthSendMfaCodeRequestDto": {
        "type": "object",
        "properties": {
          "method": {
            "description": "MFA delivery method",
            "example": "email",
            "examples": {
              "email": {
                "value": "email",
                "description": "Send OTP via email"
              },
              "sms": {
                "value": "sms",
                "description": "Send OTP via SMS"
              },
              "totp": {
                "value": "totp",
                "description": "Use authenticator app (TOTP)"
              }
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/NestAuthMFAMethodEnum"
              }
            ]
          }
        },
        "required": [
          "method"
        ]
      },
      "NestAuthVerifyTotpSetupRequestDto": {
        "type": "object",
        "properties": {
          "otp": {
            "type": "string",
            "description": "The TOTP code from authenticator app",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          },
          "secret": {
            "type": "string",
            "description": "Secret key from TOTP setup",
            "example": "JBSWY3DPEHPK3PXP"
          }
        },
        "required": [
          "otp",
          "secret"
        ]
      },
      "NestAuthMfaDeviceVerifiedResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Device setup successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}